GameObject && Player Inheritance
Working version of inheritance for Player only currently
